requirements.txt
dask
databricks==0.2
distributed
docker_py==1.10.6
h5py
numpy
pandas==1.5.1
polars==0.16.13
psutil==5.9.3
pydantic==1.10.2
pydantic_factories==1.17.2
pyspark==3.1.3
ray==2.1.0
unidist==0.2.2
geographiclib==2.0
geopy==2.3.0
haversine==2.8.0
ansicolors